Elo computation for wines

+++ web_scraper.py
@@ -10,6 +10,7 @@ from bs4 import BeautifulSoup
 import json
 import time
 import random
+import os
 
 import pickle
 from selenium.webdriver.common.proxy import Proxy, ProxyType
@@ -18,7 +19,7 @@ from selenium.webdriver.common.proxy import Proxy, ProxyType
 def set_driver_settings(proxy=False):
     options = webdriver.ChromeOptions()
     options.add_argument('--ignore-certificate-errors')
-    options.add_argument('--incognito')
+    # options.add_argument('--incognito')
 
     return options
 
@@ -142,6 +143,7 @@ def grab_review_data(review):
 
 
 def grab_user_id(soup):
+
     # grab the user name and ID
     user_info = soup.find(
         'div', class_='user-header__image-container__wrapper').find('div')['data-react-props']
@@ -155,33 +157,42 @@ def mine_review_data(user_link, driver):
 
     driver.get(user_link)
     time.sleep(5)
-    expand_show_more(driver)
 
     page_source = driver.page_source
     soup = BeautifulSoup(page_source, 'lxml')
 
-    reviews_selector = soup.find_all('div', class_='user-activity-item')
-    if len(reviews_selector) >= 1:
-        all_review_info = []
-        for r in reviews_selector:
-            review_info = grab_review_data(r)
-            all_review_info.append(review_info)
+    user_id = grab_user_id(soup)
+    existing_jsons = os.listdir('raw_data/')
+    existing_user_files = [o.split('.')[0] for o in existing_jsons]
 
-        user_id = grab_user_id(soup)
+    if str(user_id) in existing_user_files:
+        print('already scraped this person')
+        pass
+    else:
+        expand_show_more(driver)
+        page_source = driver.page_source
+        soup = BeautifulSoup(page_source, 'lxml')
+        reviews_selector = soup.find_all('div', class_='user-activity-item')
+        if len(reviews_selector) >= 1:
+            all_review_info = []
+            for r in reviews_selector:
+                review_info = grab_review_data(r)
+                all_review_info.append(review_info)
 
-        # write the scraped data to a json file
-        filename = 'raw_data/' + str(user_id) + '.json'
-        with open(filename, 'w') as outfile:
-            json.dump(all_review_info, outfile)
+            # write the scraped data to a json file
+            filename = 'raw_data/' + str(user_id) + '.json'
+            with open(filename, 'w') as outfile:
+                json.dump(all_review_info, outfile)
 
 
 def main():
     driver = get_driver(proxy=False)
-    # driver.get('https://www.vivino.com/users/martijn-kra/rankings')
+    # driver.get('https://www.vivino.com/users/roald.schuring/rankings')
 
     # time.sleep(30)
+    # # use this time to log in manually - the list of all user links is not visible if not logged in
 
-    # for _ in range(100):
+    # for _ in range(1000):
     #     show_more_button = driver.find_elements_by_class_name(
     #         'text-block.text-center.country-rankings-show-more.semi')[-1]
     #     driver.execute_script("arguments[0].click();", show_more_button)
@@ -199,9 +210,10 @@ def main():
     with open("user_links.json", 'r') as f:
         user_links = json.load(f)
 
-    for u in user_links[249:500]:
+    unique_user_links = list(set(user_links))
+    for u in unique_user_links[2600:]:
         mine_review_data(u, driver)
-        time.sleep(random.uniform(10, 30))
+        time.sleep(random.uniform(0, 10))
 
     driver.close()
